+++ src/main/java/com/cpi/is/dao/DispatchingDAO.java
@@ -6,8 +6,8 @@ import com.cpi.is.entity.DispatchingEntity;
 
 public interface DispatchingDAO {
     
-    List<DispatchingEntity> getDispatchingByBranchId(Integer branchId) throws Exception; // Add this method
-    List<Object[]> getCurrentInventory(Integer branchId) throws Exception;
+    List<DispatchingEntity> getDispatchingByBranchId(Long branchId) throws Exception; // Add this method
+    List<Object[]> getCurrentInventory(Long branchId) throws Exception;
     String saveItem(DispatchingEntity item) throws Exception;
     String deleteItem(DispatchingEntity item) throws Exception;
