+++ src/main/java/com/cpi/is/dao/inventory/RawMaterialListDAO.java
@@ -6,7 +6,7 @@ import com.cpi.is.entity.inventory.RawMaterialListEntity;
 
 public interface RawMaterialListDAO {
 	
-	List<RawMaterialListEntity>  getRawMaterialList(Integer targetBranchId) throws Exception;
+	List<RawMaterialListEntity>  getRawMaterialList(Long targetBranchId) throws Exception;
 	String saveRawMaterial(RawMaterialListEntity item) throws Exception;
 	String deleteRawMaterial(RawMaterialListEntity item) throws Exception;
 }
