+++ src/main/java/com/cpi/is/service/DispatchingService.java
@@ -9,8 +9,8 @@ import com.cpi.is.entity.inventory.FinishedProductListEntity;
 
 public interface DispatchingService {
     
-	List<DispatchingEntity> getDispatchingByBranchId(Integer branchId) throws Exception; // Add this method
-	List<Object[]> getCurrentInventory(Integer branchId) throws Exception;
+	List<DispatchingEntity> getDispatchingByBranchId(Long branchId) throws Exception; // Add this method
+	List<Object[]> getCurrentInventory(Long branchId) throws Exception;
     String deleteItem(HttpServletRequest request) throws Exception;
 	String saveItem(HttpServletRequest request, List<FinishedProductListEntity> finishedProductList) throws Exception;
 }
