+++ src/main/java/com/cpi/is/service/inventory/FinishedProductListService.java
@@ -9,7 +9,7 @@ import com.cpi.is.entity.inventory.FinishedProductListEntity;
 
 public interface FinishedProductListService {
 	
-	List<FinishedProductListEntity> getFinishedProductList(Integer branchId) throws Exception;
+	List<FinishedProductListEntity> getFinishedProductList(Long branchId) throws Exception;
 	String saveProduct(HttpServletRequest request) throws Exception;
 	String deleteProduct(HttpServletRequest request) throws Exception;
 }
